requirements.txt
requests
nltk
tweepy
pyyaml
